Moving liquid schemas to utils dir
