+++ packages/pls-liquid/index.ts
@@ -1,6 +1,3 @@
-import { PubkeysSchema } from "pls-core";
-import { z } from "zod";
-
 import { createLiquidMultisig } from "./createLiquidMultisig.js";
 import { getUnblindedUtxoValue } from "./getUnblindedUtxoValue.js";
 import { getUnblindedUtxoValues } from "./getUnblindedUtxoValues.js";
@@ -11,6 +8,7 @@ import { getTapscriptSigsOrdered } from "./getTapscriptSigsOrdered.js";
 
 import {
 	H,
+	liquidSchemas,
 } from "./utils/index.js"
 
 export {
@@ -29,23 +27,5 @@ export {
 	getTapscriptSigsOrdered,
 
 	H,
+	liquidSchemas,
 }
-
-const TaprootV0CollateralSchema = {
-	arbitratorsQuorum: z.number(),
-	multisigAddress: z.string(),
-	privateBlindingKey: z.string(),
-	pubkeys: PubkeysSchema,
-	type: z.literal("taproot-v0"),
-};
-
-export const liquidSchemas = {
-	mainnet: z.object({
-		network: z.literal("liquid"),
-		...TaprootV0CollateralSchema,
-	}),
-	testnet: z.object({
-		network: z.literal("liquid_testnet"),
-		...TaprootV0CollateralSchema,
-	}),
-};
