+++ packages/pls-liquid/utils/index.ts
@@ -1,6 +1,27 @@
+import { PubkeysSchema } from "pls-core";
+import { z } from "zod";
 import { bip341, crypto } from "liquidjs-lib";
 import secp256k1 from "@vulpemventures/secp256k1-zkp";
 import type { Secp256k1Interface } from "liquidjs-lib";
+
+const TaprootV0CollateralSchema = {
+	arbitratorsQuorum: z.number(),
+	multisigAddress: z.string(),
+	privateBlindingKey: z.string(),
+	pubkeys: PubkeysSchema,
+	type: z.literal("taproot-v0"),
+};
+
+export const liquidSchemas = {
+	mainnet: z.object({
+		network: z.literal("liquid"),
+		...TaprootV0CollateralSchema,
+	}),
+	testnet: z.object({
+		network: z.literal("liquid_testnet"),
+		...TaprootV0CollateralSchema,
+	}),
+};
 
 export async function getZkpLib(): Promise<Secp256k1Interface> {
 	// @ts-ignore I have no idea
